requirements.txt
Django==1.6.1
PIL==1.1.7
Pillow==2.3.0
South==0.8.4
Unipath==1.0
amqp==1.3.3
anyjson==0.3.3
beautifulsoup4==4.3.2
billiard==3.3.0.11
boto==2.21.2
celery==3.1.6
diff-match-patch==20121119
dj-database-url==0.2.2
django-nose==1.2
django-redis-cache==0.10.2
django-reversion==1.8.0
django-storages==1.1.8
gunicorn==18.0
hiredis==0.1.2
honcho==0.5.0
kombu==3.0.7
lxml==3.2.4
nose==1.3.0
psycopg2==2.5.1
pytz==2013.8
redis==2.8.0
requests==2.1.0
sorl-thumbnail==11.12
wsgiref==0.1.2
